+++ examples/UsingAbstractClass/Consumers/Consumer1.cs
@@ -1,5 +1,4 @@
-﻿using Confluent.Kafka;
-using Reactive.Kafka;
+﻿using Reactive.Kafka;
 
 namespace UsingAbstractClass.Consumers
 {
@@ -8,18 +7,14 @@ namespace UsingAbstractClass.Consumers
         private readonly ILogger<Consumer1> _logger;
 
         public Consumer1(ILogger<Consumer1> logger)
-            => _logger = logger;
+        {
+            _logger = logger;
+        }
 
         public override Task OnConsume(ConsumerMessage<Message> consumerMessage, Commit commit)
         {
-            _logger.LogInformation("Message ==> {Message}", consumerMessage.Message);
-
+            _logger.LogInformation("{Message}", consumerMessage.Message);
             return Task.CompletedTask;
-        }
-
-        public override void OnConsumerConfiguration(IConsumer<string, string> consumer)
-        {
-            consumer.Subscribe("your-topic");
         }
     }
 }