+++ src/Exceptions/KafkaConsumerException.cs
@@ -2,6 +2,12 @@
 {
     public class KafkaConsumerException : Exception
     {
-        public KafkaConsumerException(string message) : base(message) { }
+        public KafkaConsumerException(string message, string messageKafka)
+            : base(message)
+        {
+            MessageKafka = messageKafka;
+        }
+
+        public string MessageKafka { get; }
     }
 }