+++ src/Exceptions/KafkaSerializationException.cs
@@ -1,0 +1,8 @@
+﻿namespace Reactive.Kafka.Exceptions
+{
+    public class KafkaSerializationException : KafkaConsumerException
+    {
+        public KafkaSerializationException(string message, string messageKafka)
+            : base(message, messageKafka) { }
+    }
+}